refactor(payments): Move helpers under lib/payments
Keep the public payments:* Convex paths stable, but pull Prava client,
money, mandate, and charge helpers into focused modules so the money
paths are easier to review.
