+++ apps/web/src/convex/payments.ts
@@ -2,24 +2,18 @@ import { ActionRetrier, onCompleteValidator, type RunId } from '@convex-dev/acti
 import { v, type Infer, type ObjectType } from 'convex/values';
 import {
 	action,
-	env,
 	internalAction,
 	internalMutation,
 	internalQuery,
-	type ActionCtx,
-	type MutationCtx
+	type ActionCtx
 } from '@convex/_generated/server';
-import { api, components, internal } from '@convex/_generated/api';
-import type { Doc, Id } from '@convex/_generated/dataModel';
+import { components, internal } from '@convex/_generated/api';
 import { getUserId, pickPrimaryUser } from '@convex/lib/auth';
-import { isRunClaimLeaseActive } from '@convex/lib/runLease';
 import { toAgentToolConvexError } from '@convex/lib/agentErrors';
 import { unsupportedClient } from '@convex/lib/unsupportedClient';
 import {
-	isMandateStatus,
-	vMandateChargeStatus,
-	vMandateFrequency,
 	vMandateReportOutcome,
+	vMandateFrequency,
 	vMandateScope,
 	vMandateStatus,
 	vMandateSetupResult,
@@ -28,160 +22,37 @@ import {
 	vMandateChargeResult,
 	vMandateReportResult
 } from '@convex/lib/validators';
-
-const REPORT_CLAIM_STALE_MS = 60_000;
-const CHARGE_CLAIM_STALE_MS = 60_000;
+import { requireMoneyMinor } from '@convex/lib/payments/money';
+import { pravaRequest, type PravaMandate } from '@convex/lib/payments/prava';
+import {
+	activeActor,
+	assertMandateFrequencyAllowed,
+	LIVE_MANDATE_STATUSES,
+	listLinkedMandates,
+	mandateDoc,
+	mandateStatusResult,
+	mandateSyncArgs,
+	resolvePravaMandate,
+	withMandateSync
+} from '@convex/lib/payments/mandates';
+import {
+	assertChargeable,
+	CHARGE_CLAIM_STALE_MS,
+	chargeDoc,
+	type ChargeReportPostArgs,
+	type ChargeReportReleasePatch,
+	type ChargeStatusPatch,
+	type MandateChargeRequest,
+	type MandateReportRequest,
+	ownedCharge,
+	REPORT_CLAIM_STALE_MS,
+	statusForOutcome
+} from '@convex/lib/payments/charges';
 
 const reportRetrier = new ActionRetrier(components.actionRetrier, {
 	initialBackoffMs: 250,
 	base: 2,
 	maxFailures: 4
-});
-
-type PravaConfig = {
-	baseUrl: string;
-	secretKey: string;
-};
-
-function pravaConfig(): PravaConfig {
-	const secretKey = env.PRAVA_SECRET_KEY?.trim();
-	if (!secretKey) {
-		throw new Error('PRAVA_SECRET_KEY is not configured.');
-	}
-	return {
-		baseUrl: env.PRAVA_BACKEND_URL,
-		secretKey
-	};
-}
-
-async function pravaRequest<T>(path: string, init?: RequestInit): Promise<T> {
-	const { baseUrl, secretKey } = pravaConfig();
-	const headers = new Headers(init?.headers);
-	headers.set('Authorization', `Bearer ${secretKey}`);
-	if (init?.body) headers.set('Content-Type', 'application/json');
-	const response = await fetch(`${baseUrl}${path}`, {
-		...init,
-		headers
-	});
-	if (!response.ok) {
-		const details = await response.text();
-		let message = details;
-		try {
-			// SAFETY: Prava errors share one documented envelope
-			// ({error:{code,message,details}}); both fields are optional-checked
-			// before use and anything else keeps the raw body as the message.
-			const parsed = JSON.parse(details) as { error?: { code?: string; message?: string } };
-			if (parsed.error?.code || parsed.error?.message) {
-				message = [parsed.error.code, parsed.error.message].filter(Boolean).join(' - ');
-			}
-		} catch {
-			// Not JSON; surface the raw body.
-		}
-		throw new Error(`Prava request failed (${response.status})${message ? `: ${message}` : '.'}`);
-	}
-	const body = await response.text();
-	// SAFETY: unchecked decode of the trusted Prava API response into its documented contract T.
-	return (body ? JSON.parse(body) : undefined) as T;
-}
-
-/** A first-time customer has no Prava customer record yet, so listing their
- * mandates 404s with CUSTOMER_NOT_FOUND; treat that as "no mandates yet"
- * rather than an error. */
-function isPravaCustomerNotFound(error: Error): boolean {
-	return error.message.includes('CUSTOMER_NOT_FOUND');
-}
-
-/** List a customer's mandates from Prava, treating a first-time customer's
- * CUSTOMER_NOT_FOUND 404 as an empty list rather than an error. */
-async function listPravaMandates(userId: string, standingOnly: boolean): Promise<PravaMandate[]> {
-	try {
-		const list = await pravaRequest<{ mandates?: PravaMandate[] }>(
-			`/v1/mandates?customer_id=${encodeURIComponent(userId)}${standingOnly ? '&standing_only=true' : ''}`
-		);
-		return list.mandates ?? [];
-	} catch (error) {
-		if (error instanceof Error && isPravaCustomerNotFound(error)) {
-			return [];
-		}
-		throw error;
-	}
-}
-
-/** Only these mandate states can be charged or should be surfaced to the
- * agent/user. Cancelled/expired/consumed approvals must not match a local
- * setup during resolution. A stale same-merchant+amount approval would
- * otherwise poison the unique-match check forever. */
-const LIVE_MANDATE_STATUSES = new Set(['pending', 'active', 'paused']);
-
-async function activeActor(
-	ctx: ActionCtx,
-	args: { runId: Doc<'runs'>['_id']; claimId: string; executionSecret: string }
-) {
-	const actor = await ctx.runQuery(api.agentRuntime.completionActor, {
-		runId: args.runId,
-		executionSecret: args.executionSecret
-	});
-	if (actor.claimId !== args.claimId || !isRunClaimLeaseActive(actor, Date.now())) {
-		throw new Error('Run is no longer active.');
-	}
-	return actor;
-}
-
-/** Any-scope (generic) mandates are one-time only. Prava rejects a recurring
- * frequency with any-scope, but enforce it locally so a one-time authorization
- * can never become a reusable recurring one. */
-function assertMandateFrequencyAllowed(args: {
-	scope: Infer<typeof vMandateScope>;
-	frequency: Infer<typeof vMandateFrequency>;
-}) {
-	if (args.scope === 'any' && args.frequency !== 'one_time') {
-		throw new Error('Any-merchant mandates must be one-time.');
-	}
-}
-
-const mandateDoc = v.object({
-	_id: v.id('mandates'),
-	_creationTime: v.number(),
-	userId: v.string(),
-	pravaMandateId: v.optional(v.string()),
-	pravaSessionId: v.string(),
-	merchantName: v.optional(v.string()),
-	merchantUrl: v.optional(v.string()),
-	countryCode: v.optional(v.string()),
-	amountCap: v.number(),
-	currency: v.string(),
-	frequency: vMandateFrequency,
-	scope: vMandateScope,
-	status: vMandateStatus,
-	description: v.string(),
-	approvalUrl: v.string(),
-	validUntil: v.optional(v.string()),
-	renewsAt: v.optional(v.string()),
-	remaining: v.optional(v.number()),
-	createdAt: v.number(),
-	updatedAt: v.number()
-});
-
-const chargeDoc = v.object({
-	_id: v.id('mandateCharges'),
-	_creationTime: v.number(),
-	mandateId: v.id('mandates'),
-	runId: v.id('runs'),
-	userId: v.string(),
-	pravaTransactionId: v.optional(v.string()),
-	amount: v.number(),
-	currency: v.string(),
-	description: v.string(),
-	reference: v.optional(v.string()),
-	status: vMandateChargeStatus,
-	reportOutcome: v.optional(vMandateReportOutcome),
-	reportedAt: v.optional(v.number()),
-	reportingStartedAt: v.optional(v.number()),
-	chargingStartedAt: v.optional(v.number()),
-	providerRequestedAt: v.optional(v.number()),
-	reportRetrierRunId: v.optional(v.string()),
-	createdAt: v.number(),
-	updatedAt: v.number()
 });
 
 type MandateSyncPatch = {
@@ -193,27 +64,6 @@ type MandateSyncPatch = {
 	renewsAt?: string;
 };
 
-type ChargeStatusPatch = {
-	status: 'completed' | 'declined' | 'failed';
-	chargingStartedAt: undefined;
-	updatedAt: number;
-	providerRequestedAt?: undefined;
-};
-
-type ChargeReportPostArgs = {
-	chargeId: Id<'mandateCharges'>;
-	userId: string;
-	outcome: Infer<typeof vMandateReportOutcome>;
-	amountPaid?: string;
-};
-
-type ChargeReportReleasePatch = {
-	reportingStartedAt: undefined;
-	updatedAt: number;
-	reportOutcome?: undefined;
-	reportRetrierRunId?: undefined;
-};
-
 type MandateSetupRequest = {
 	intent: 'mandate_setup';
 	recurring_frequency: Infer<typeof vMandateFrequency>;
@@ -221,151 +71,6 @@ type MandateSetupRequest = {
 	max_charges?: number;
 	valid_until?: string;
 };
-
-type MandateChargeRequest = {
-	amount: string;
-	reference?: string;
-};
-
-type MandateReportRequest = {
-	txn_status: 'APPROVED' | 'DECLINED';
-	txn_type: 'PURCHASE';
-	amount_paid?: string;
-};
-
-type PravaMandate = {
-	id?: string;
-	status?: string;
-	recurringFrequency?: string;
-	merchantScope?: string;
-	remaining?: string | null;
-	approvedAmount?: string;
-	currency?: string;
-	merchantName?: string | null;
-	validUntil?: string | null;
-	renewsAt?: string | null;
-};
-
-function pravaStatusToLocal(status: string | undefined): Infer<typeof vMandateStatus> | undefined {
-	return isMandateStatus(status) ? status : undefined;
-}
-
-/** Parse a Prava/agent decimal money string into integer minor units (cents).
- * Fixed-point, so "0.1" + "0.2" class errors can't leak into comparisons. */
-function parseMoneyMinor(value: string): number | undefined {
-	const match = /^(\d+)(?:\.(\d{1,2}))?$/.exec(value.trim());
-	if (!match) return undefined;
-	const units = Number(match[1]);
-	const fraction = Number((match[2] ?? '').padEnd(2, '0') || '0');
-	if (!Number.isSafeInteger(units) || !Number.isSafeInteger(fraction)) return undefined;
-	const minor = units * 100 + fraction;
-	return Number.isSafeInteger(minor) ? minor : undefined;
-}
-
-function requireMoneyMinor(value: string, label: string): number {
-	const minor = parseMoneyMinor(value);
-	if (minor === undefined) {
-		throw new Error(`${label} must be a non-negative decimal amount.`);
-	}
-	return minor;
-}
-
-/** Format integer minor units for Prava/agent/UI decimal strings. Minor units
- * only enter through parseMoneyMinor, so they are always non-negative. */
-function formatMoneyMinor(minor: number): string {
-	return `${Math.floor(minor / 100)}.${(minor % 100).toString().padStart(2, '0')}`;
-}
-
-/** Fail fast on a charge the mandate obviously can't authorize. Prava is
- * still the authoritative limit enforcer at the card network. These checks
- * only stop clearly wrong requests from producing misleading local records. */
-function assertChargeable(
-	mandate: Doc<'mandates'>,
-	args: { amount: string; currency: string }
-): void {
-	if (args.currency.trim().toUpperCase() !== mandate.currency.toUpperCase()) {
-		throw new Error(`Charge currency must match the mandate's ${mandate.currency}.`);
-	}
-	const amount = parseMoneyMinor(args.amount);
-	if (amount === undefined || amount <= 0) {
-		throw new Error('Charge amount must be a positive decimal string.');
-	}
-	if (amount > mandate.amountCap) {
-		throw new Error(
-			`Charge amount exceeds the mandate's ${formatMoneyMinor(mandate.amountCap)} cap.`
-		);
-	}
-	if (mandate.remaining !== undefined && amount > mandate.remaining) {
-		throw new Error(
-			`Charge amount exceeds the mandate's remaining ${formatMoneyMinor(mandate.remaining)}.`
-		);
-	}
-}
-
-async function ownedCharge(
-	ctx: MutationCtx,
-	chargeId: Id<'mandateCharges'>,
-	userId: string
-): Promise<Doc<'mandateCharges'>> {
-	const charge = await ctx.db.get('mandateCharges', chargeId);
-	if (!charge || charge.userId !== userId) {
-		throw new Error('Charge not found.');
-	}
-	return charge;
-}
-
-function statusForOutcome(outcome: Infer<typeof vMandateReportOutcome>): 'completed' | 'declined' {
-	return outcome === 'approved' ? 'completed' : 'declined';
-}
-
-function mandateSyncArgs(
-	mandateId: Id<'mandates'>,
-	userId: string,
-	mandate: PravaMandate & { id: string }
-) {
-	return {
-		mandateId,
-		userId,
-		pravaMandateId: mandate.id,
-		status: pravaStatusToLocal(mandate.status),
-		remaining: mandate.remaining == null ? undefined : parseMoneyMinor(mandate.remaining),
-		validUntil: mandate.validUntil ?? undefined,
-		renewsAt: mandate.renewsAt ?? undefined
-	};
-}
-
-/** Overlay freshly synced Prava fields onto the local mandate doc. */
-function withMandateSync(
-	mandate: Doc<'mandates'>,
-	sync: ReturnType<typeof mandateSyncArgs>
-): Doc<'mandates'> {
-	return {
-		...mandate,
-		pravaMandateId: sync.pravaMandateId,
-		status: sync.status ?? mandate.status,
-		remaining: sync.remaining ?? mandate.remaining,
-		validUntil: sync.validUntil ?? mandate.validUntil,
-		renewsAt: sync.renewsAt ?? mandate.renewsAt
-	};
-}
-
-function mandateStatusResult(mandate: Doc<'mandates'>): Infer<typeof vMandateStatusResult> {
-	return {
-		mandateId: mandate._id,
-		pravaMandateId: mandate.pravaMandateId,
-		status: mandate.status,
-		description: mandate.description,
-		merchantName: mandate.merchantName,
-		amountCap: formatMoneyMinor(mandate.amountCap),
-		remaining: mandate.remaining !== undefined ? formatMoneyMinor(mandate.remaining) : undefined,
-		currency: mandate.currency,
-		frequency: mandate.frequency,
-		scope: mandate.scope,
-		approvalUrl: mandate.approvalUrl,
-		validUntil: mandate.validUntil,
-		renewsAt: mandate.renewsAt
-	};
-}
 
 // ---------------------------------------------------------------------------
 // Internal queries / mutations
@@ -1019,180 +724,6 @@ export const mandateSetup = action({
 		}
 	}
 });
-
-/** True when a live Prava mandate uniquely matches a local setup's merchant,
- * amount cap, and currency. Used both for charge-time resolution and for
- * linking local rows after the owner approves in a new tab. */
-function isMatchingLivePravaMandate(mandate: Doc<'mandates'>, prava: PravaMandate): boolean {
-	if (!prava.id) return false;
-	if (!LIVE_MANDATE_STATUSES.has(prava.status ?? '')) return false;
-	// A different scope or cadence is a different authorization even when the
-	// merchant name and cap coincide.
-	if (prava.merchantScope !== undefined && prava.merchantScope !== mandate.scope) return false;
-	if (prava.recurringFrequency !== undefined && prava.recurringFrequency !== mandate.frequency) {
-		return false;
-	}
-	// A mandate approved in another currency is a different authorization;
-	// never resolve (and later charge) it for this setup.
-	if ((prava.currency ?? '').toUpperCase() !== mandate.currency.toUpperCase()) return false;
-	const approvedMinor = prava.approvedAmount ? parseMoneyMinor(prava.approvedAmount) : undefined;
-	if (approvedMinor === undefined || approvedMinor !== mandate.amountCap) return false;
-	if (mandate.scope === 'listed') {
-		const localName = mandate.merchantName?.trim();
-		if (!localName) return false;
-		return (prava.merchantName ?? '').toLowerCase() === localName.toLowerCase();
-	}
-	return true;
-}
-
-function matchingLivePravaMandates(
-	mandate: Doc<'mandates'>,
-	list: PravaMandate[],
-	claimedIds?: ReadonlySet<string>
-): Array<PravaMandate & { id: string }> {
-	return list
-		.filter((m) => {
-			if (!m.id || claimedIds?.has(m.id)) return false;
-			return isMatchingLivePravaMandate(mandate, m);
-		})
-		.map((m) => ({ ...m, id: m.id! }));
-}
-
-function isUnresolvedLocal(mandate: Doc<'mandates'>): boolean {
-	return !mandate.pravaMandateId && LIVE_MANDATE_STATUSES.has(mandate.status);
-}
-
-/** A Prava mandate is only safe to bind when it uniquely matches this local
- * setup and no other unresolved local setup also matches it. Otherwise two
- * same-merchant/amount pending rows can steal each other's approval. */
-function uniquelyAttributablePravaMandate(
-	mandate: Doc<'mandates'>,
-	list: PravaMandate[],
-	allLocal: Doc<'mandates'>[]
-):
-	| { kind: 'matched'; mandate: PravaMandate & { id: string } }
-	| { kind: 'none' }
-	| { kind: 'ambiguous' } {
-	const claimed = new Set(
-		allLocal
-			.filter((row) => row.pravaMandateId && row._id !== mandate._id)
-			.map((row) => row.pravaMandateId)
-			.filter((id): id is string => id !== undefined)
-	);
-	const matches = matchingLivePravaMandates(mandate, list, claimed);
-	if (matches.length === 0) return { kind: 'none' };
-	if (matches.length > 1) return { kind: 'ambiguous' };
-	const candidate = matches[0];
-	const contested = allLocal.some(
-		(peer) =>
-			peer._id !== mandate._id &&
-			isUnresolvedLocal(peer) &&
-			isMatchingLivePravaMandate(peer, candidate)
-	);
-	if (contested) return { kind: 'ambiguous' };
-	return { kind: 'matched', mandate: candidate };
-}
-
-/** Resolve the Prava-side mandate id for a locally stored mandate. It only
- * exists once the owner approves, so resolve lazily by listing the user's
- * mandates and matching this setup's merchant + cap, then persist it. */
-async function resolvePravaMandate(
-	ctx: ActionCtx,
-	userId: string,
-	mandate: Doc<'mandates'>
-): Promise<PravaMandate & { id: string }> {
-	if (mandate.pravaMandateId) {
-		const found = await pravaRequest<PravaMandate>(
-			`/v1/mandates/${encodeURIComponent(mandate.pravaMandateId)}`
-		);
-		return { ...found, id: mandate.pravaMandateId };
-	}
-	const list = await listPravaMandates(userId, false);
-	const local = await ctx.runQuery(internal.payments.listLocalMandates, { userId });
-	const match = uniquelyAttributablePravaMandate(mandate, list, local);
-	// Require a unique live match. Charging an arbitrary same-merchant+amount
-	// approval could settle against the wrong authorization. Prava mandates
-	// don't carry their setup session id, so there is no stronger link to
-	// prefer on.
-	if (match.kind === 'none') {
-		throw new Error('Mandate is not yet approved.');
-	}
-	if (match.kind === 'ambiguous') {
-		throw new Error(
-			'Cannot uniquely match this setup to an approved mandate; resolve the ambiguity before charging.'
-		);
-	}
-	const resolved = match.mandate;
-	await ctx.runMutation(
-		internal.payments.syncMandate,
-		mandateSyncArgs(mandate._id, userId, resolved)
-	);
-	return resolved;
-}
-
-/** Join Prava's live mandates to local rows. After new-tab approval the local
- * row still lacks pravaMandateId, so uniquely match unresolved locals against
- * the live list and persist the link; otherwise settings can't pause/cancel. */
-async function linkLocalMandates(
-	ctx: ActionCtx,
-	userId: string,
-	list: PravaMandate[]
-): Promise<{
-	localByPravaId: Map<string, Id<'mandates'>>;
-	localById: Map<Id<'mandates'>, Doc<'mandates'>>;
-}> {
-	const local = await ctx.runQuery(internal.payments.listLocalMandates, { userId });
-	const localById = new Map(local.map((m) => [m._id, m]));
-	const localByPravaId = new Map(
-		local
-			.filter((m): m is typeof m & { pravaMandateId: string } => m.pravaMandateId !== undefined)
-			.map((m) => [m.pravaMandateId, m._id])
-	);
-	for (const mandate of local.filter(isUnresolvedLocal)) {
-		const match = uniquelyAttributablePravaMandate(mandate, list, local);
-		if (match.kind !== 'matched') continue;
-		const sync = mandateSyncArgs(mandate._id, userId, match.mandate);
-		await ctx.runMutation(internal.payments.syncMandate, sync);
-		const linked = withMandateSync(mandate, sync);
-		localByPravaId.set(sync.pravaMandateId, mandate._id);
-		localById.set(mandate._id, linked);
-		// Keep later uniqueness checks aware of the link we just persisted.
-		local[local.indexOf(mandate)] = linked;
-	}
-	return { localByPravaId, localById };
-}
-
-/** List the user's live Prava mandates joined to local rows (linking any
- * still-unresolved setups) so each entry carries the local mandateId the
- * lifecycle action needs. Entries without a local row (e.g. approved
- * elsewhere) are still listed, just without a mandateId. */
-async function listLinkedMandates(
-	ctx: ActionCtx,
-	userId: string
-): Promise<Infer<typeof vMandateListResult>> {
-	const list = (await listPravaMandates(userId, false)).filter((m) =>
-		LIVE_MANDATE_STATUSES.has(m.status ?? '')
-	);
-	const { localByPravaId, localById } = await linkLocalMandates(ctx, userId, list);
-	return {
-		mandates: list.map((m) => {
-			const mandateId = m.id ? localByPravaId.get(m.id) : undefined;
-			const local = mandateId ? localById.get(mandateId) : undefined;
-			return {
-				mandateId,
-				pravaMandateId: m.id ?? '',
-				status: m.status ?? 'unknown',
-				description: local?.description,
-				merchantName: m.merchantName ?? local?.merchantName ?? undefined,
-				approvedAmount: m.approvedAmount ?? '',
-				remaining: m.remaining ?? undefined,
-				currency: m.currency ?? '',
-				validUntil: m.validUntil ?? undefined,
-				renewsAt: m.renewsAt ?? undefined
-			};
-		})
-	};
-}
 
 export const mandateStatus = action({
 	args: {
